+++ src/icons/animals/index.ts
@@ -1,0 +1,3 @@
+import DogPaw from '@/src/icons/animals/DogPaw';
+
+export { DogPaw };