+++ src/icons/transportation/index.ts
@@ -4,5 +4,23 @@
 import Airplane from '@/src/icons/transportation/Airplane';
 import Bus from '@/src/icons/transportation/Bus';
 import Cab from '@/src/icons/transportation/Cab';
+import Car1 from '@/src/icons/transportation/Car1';
+import Car2 from '@/src/icons/transportation/Car2';
+import CarService from '@/src/icons/transportation/CarService';
+import FuelDispenserNozzle from '@/src/icons/transportation/FuelDispenserNozzle';
+import Parking from '@/src/icons/transportation/Parking';
+import Train from '@/src/icons/transportation/Train';
+import Tram from '@/src/icons/transportation/Tram';
 
-export { Airplane, Bus, Cab };
+export {
+    Airplane,
+    Bus,
+    Cab,
+    Car1,
+    Car2,
+    CarService,
+    FuelDispenserNozzle,
+    Parking,
+    Train,
+    Tram
+};